+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- updated to 1.18.2
+- switched to Fabric